+++ loki/loki_test.go
@@ -592,7 +592,7 @@ func TestOutputFactory_LoggerReachesOutput(t *testing.T) {
 	yaml := []byte(
 		"url: https://loki.example.com/loki/api/v1/push\n" +
 			"verify_on_startup: false\n" +
-			"tls_policy:\n" +
+			"tls:\n" +
 			"  allow_tls12: true\n" +
 			"  allow_weak_ciphers: true\n" +
 			"batch_size: 1\nflush_interval: 1s\ntimeout: 5s\n",